Modify deletion function and add reverse function
During the execution of q_delete_dup, the code identified a problem
with blocks still being allocated.
To address this, I made modifications to ensure that the memory allocated
for the original 'prob' is freed before updating its memory location.
This resolves the issue of memory blocks not being properly deallocated.

Implement q_reverseK function to reverse the nodes of the list k at a time.
The function calculates the length of the queue and then performs a partial
reverse by swapping the prev and next pointers of each element. Additionally,
adjustments are made to the head and tail nodes to ensure proper reversal of
the segments.